+++ assets/project-overview.js
@@ -42,14 +42,24 @@ $(document).ready(function () {
       });
 
     // Assign event handlers to dynamic elements
-    $(document).on('click', '.dropdown-item .table-button', function () {
+    $(document).on('click', '.dropdown-item .table-button.status', function () {
       const idArgs = $(this).data('args').split(',');
       changeStatus(idArgs[0], idArgs[1], idArgs[2], idArgs[3]);
     });
 
+    $(document).on(
+      'click',
+      '.dropdown-item .table-button.priority',
+      function () {
+        const idArgs = $(this).data('args').split(',');
+        changePriority(idArgs[0], idArgs[1], idArgs[2]);
+      }
+    );
+
     $(document).on('change', '[id^=due-date-]', function () {
-      const idArg = $(this).attr('id').split('-')[2];
-      changeDueDate(event, idArg, this.value);
+      const date = $(this).val();
+      const ticketId = $(this).data('ticketid');
+      changeDueDate(event, ticketId, date);
     });
 
     $(document).on('change', '[id^=assigned-user-]', function () {
@@ -94,7 +104,7 @@ $(document).ready(function () {
       changeDateFrom(this.value);
     });
 
-    $('#date-to').on('change', function () {
+    $('#date-to').on('focusout', function () {
       changeDateTo(this.value);
     });
   });
@@ -152,7 +162,7 @@ function changeDueDate(event, ticketId, newDueDate) {
   const parentElement = jQuery(event.target).closest('td');
 
   if (newDueDate && ticketId) {
-    const dueDate = jQuery.datepicker.formatDate(
+    const dueDate = window.jQuery.datepicker.formatDate(
       leantime.dateHelper.getFormatFromSettings('dateformat', 'jquery'),
       new Date(newDueDate)
     );
@@ -241,6 +251,7 @@ function changeHoursRemaining(event, ticketId, newHoursRemaining) {
 }
 
 function changeMilestone(event, ticketId, newMilestoneId) {
+  const parentElement = jQuery(event.target).closest('td');
   if (newMilestoneId && ticketId) {
     jQuery
       .ajax({
@@ -251,27 +262,11 @@ function changeMilestone(event, ticketId, newMilestoneId) {
           milestoneid: newMilestoneId,
         },
       })
-      .done(() => {
-        // In this way, the UI does not reflect the actual data, which is not good.
-        // But if I instead create a get-request it returns 200 and an otherwise empty
-        // response. So this is what I chose to do, and is also what is done in
-        // in other places (I am looking at you ticketcontroller.js).
-        const parentElement = event.target;
-        const hexColorRegExp = new RegExp('^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$');
-        const newMilestoneColor = jQuery(
-          `#milestone-option-${newMilestoneId}`
-        ).attr('data-color');
-
-        const isItAHexColor = hexColorRegExp.exec(newMilestoneColor);
-        if (isItAHexColor) {
-          jQuery(parentElement).css('background', newMilestoneColor);
-          jQuery('#milestone-select').addClass('milestone-select-white-text');
-        } else {
-          jQuery('#milestone-select').removeClass(
-            'milestone-select-white-text'
-          );
-          jQuery(parentElement).css('background', 'transparent');
-        }
+      .then(() => {
+        saveSuccess(parentElement);
+      })
+      .fail(() => {
+        saveError(parentElement);
       });
   }
 }